[Docs] Error handling on gitstamp, formatting of timestamp
Moved date format out to conf.py so it's not hardcoded.
Made invocation of git to be global so we are not instantiating for each
page.
Added some error handling around failure to locate a git repo.
Added documentation of dependency.

Issue #2028
